requirements.txt
pytest-playwright
pytest-xdist
pytest